move all vscode-specific files to own directory
